requirements.txt
fastapi-cli==0.0.7
uvicorn
sqlalchemy
loguru
apscheduler
python-dotenv==1.1.0
requests==2.32.3
fastapi